refine hierarchy
