+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/serializer/converters/DataConverter.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.utils.serializer
+package com.intel.analytics.bigdl.utils.serializer.converters
 
 import com.google.protobuf.ByteString
 
@@ -24,9 +24,10 @@ import com.intel.analytics.bigdl.nn.abstractnn.DataFormat.{NCHW, NHWC}
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, DataFormat}
 import com.intel.analytics.bigdl.nn.quantized._
 import com.intel.analytics.bigdl.optim.{L1L2Regularizer, L1Regularizer, L2Regularizer, Regularizer}
-import com.intel.analytics.bigdl.tensor.{DenseType, QuantizedTensor, QuantizedType, Tensor, Storage}
+import com.intel.analytics.bigdl.tensor.{TensorType => _, _}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric.{NumericBoolean, NumericChar, NumericDouble, NumericFloat, NumericInt, NumericLong, NumericShort, NumericString}
+import com.intel.analytics.bigdl.utils.serializer._
 import com.intel.analytics.bigdl.utils.tf.TFTensorNumeric.NumericByteString
 import serialization.Bigdl._
 import serialization.Bigdl.AttrValue.ArrayValue
@@ -188,512 +189,6 @@ object DataConverter extends DataConverter{
       DataFormatConverter.setAttributeValue(context, attributeBuilder, value)
     } else {
       CustomConverterDelegator.setAttributeValue(context, attributeBuilder, value, valueType)
-    }
-  }
-
-/**
- * DataConverter for [[com.intel.analytics.bigdl.optim.Regularizer]]
- */
-  object RegularizerConverter extends DataConverter {
-
-    override def getAttributeValue[T : ClassTag](context: DeserializeContext,
-                                                 attribute: AttrValue)
-      (implicit ev: TensorNumeric[T]): AnyRef = {
-      val regularizer = attribute.getRegularizerValue
-      val regularizerType = regularizer.getRegularizerType
-      if (regularizer.getRegularDataCount == 0) {
-        return null
-      }
-      regularizerType match {
-        case serialization.Bigdl.RegularizerType.L1Regularizer =>
-          val l1 = regularizer.getRegularDataList.get(0)
-          L1Regularizer[T](l1)
-        case serialization.Bigdl.RegularizerType.L2Regularizer =>
-          val l2 = regularizer.getRegularDataList.get(1)
-          L2Regularizer[T](l2)
-        case serialization.Bigdl.RegularizerType.L1L2Regularizer =>
-          val l1 = regularizer.getRegularDataList.get(0)
-          val l2 = regularizer.getRegularDataList.get(1)
-          L1L2Regularizer[T](l1, l2)
-      }
-    }
-
-    override def setAttributeValue[T : ClassTag]
-    (context: SerializeContext[T], attributeBuilder: AttrValue.Builder, value: Any,
-     valueType : universe.Type = null)
-    (implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.REGULARIZER)
-      if (value != null) {
-        var regularizerBuilder = serialization.Bigdl.Regularizer.newBuilder
-        val regularizer = value.asInstanceOf[L1L2Regularizer[T]]
-        val l1 = regularizer.l1
-        val l2 = regularizer.l2
-        regularizerBuilder.addRegularData(l1)
-        regularizerBuilder.addRegularData(l2)
-        val regularizerType = regularizer match {
-          case l1: L1Regularizer[_] => serialization.Bigdl.RegularizerType.L1Regularizer
-          case l2: L2Regularizer[_] => serialization.Bigdl.RegularizerType.L2Regularizer
-          case l1l2: L1L2Regularizer[_] => serialization.Bigdl.RegularizerType.L1L2Regularizer
-        }
-        regularizerBuilder.setRegularizerType(regularizerType)
-        attributeBuilder.setRegularizerValue(regularizerBuilder.build)
-      }
-    }
-
-  }
-
-/**
- * DataConverter for [[com.intel.analytics.bigdl.tensor.Tensor]]
- */
-  object TensorConverter extends DataConverter {
-
-
-    private def isEmptyTensor(tensor : Tensor[_]): Boolean = {
-      val emptyTensor = tensor.getTensorType match {
-        case DenseType =>
-          tensor.storage == null
-        case QuantizedType =>
-          tensor.asInstanceOf[QuantizedTensor[_]].getStorage == null
-      }
-      emptyTensor
-    }
-
-    override def getAttributeValue[T: ClassTag](context: DeserializeContext,
-                                                attribute: AttrValue)
-      (implicit ev: TensorNumeric[T]): AnyRef = {
-      val serializedTensor = attribute.getTensorValue
-      if (!serializedTensor.hasStorage) {
-        return null
-      }
-      val storages = context.storages
-      val tensorId = serializedTensor.getId
-      if (storages.contains(tensorId)) {
-        return storages.get(tensorId).get.asInstanceOf[AnyRef]
-      }
-      val dataType = serializedTensor.getDatatype
-      val tensorType = serializedTensor.getTensorType
-      val sizes = serializedTensor.getSizeList.asScala.toArray.map(_.intValue())
-      val strides = serializedTensor.getStrideList.asScala.toArray.map(_.intValue())
-      val offSet = serializedTensor.getOffset
-      val isScalr = serializedTensor.getIsScalar
-      val serializedStorage = serializedTensor.getStorage
-      val storageId = serializedStorage.getId
-      val created = if (storages.contains(storageId)) {
-        storages.get(storageId).get
-      } else {
-        null
-      }
-
-      def quant(): Tensor[T] = {
-        var bytes: Array[Byte] = null
-        if (context.storageType == ProtoStorageType) {
-          bytes = serializedStorage.getBytesDataList.asScala.toArray.head.toByteArray
-        } else {
-          created
-        }
-        val serializedParams = serializedStorage.getIntDataList.asScala.toArray.map(_.intValue())
-        val paramsNum = serializedParams.head
-        val paramsArray = serializedParams.slice(1, paramsNum + 1)
-        val descTypeEnum = serializedParams(1 + paramsNum)
-
-        val start = paramsNum + 2 // params number indicator + params number + desc type
-
-        val length = if (sizes.length == 1) {
-          1 // if the size is 1, means it's a vector
-        } else {
-          sizes(0)
-        }
-        val max = new Array[T](length)
-        val min = new Array[T](length)
-        val sum = new Array[T](length)
-
-        dataType match {
-          case DataType.FLOAT =>
-            val data = serializedStorage.getFloatDataList.asScala.toArray.map(_.floatValue())
-            var i = 0
-            while (i < length) {
-              max(i) = ev.fromType[Float](data(i))
-              min(i) = ev.fromType[Float](data(i + length))
-              sum(i) = ev.fromType[Float](data(i + 2 * length))
-              i += 1
-            }
-        }
-
-        var params: DescParams = null
-
-        descTypeEnum match {
-          case 0 =>
-            params = ConvDataParams(paramsArray)
-          case 1 =>
-            params = ConvWeightParams(paramsArray)
-          case 2 =>
-            params = LinearDataParams(paramsArray)
-          case 3 =>
-            params = LinearWeightParams(paramsArray)
-        }
-
-        QuantizedTensor[T](bytes, max, min, sum, sizes, params)
-      }
-
-      val tensor = dataType match {
-        case DataType.FLOAT =>
-          tensorType match {
-            case TensorType.DENSE =>
-              val storage : Storage[Float] = if (created == null ) {
-                if (storageId == -1) {
-                  null
-                } else {
-                  val data = serializedStorage.getFloatDataList.asScala.toArray.map(_.floatValue())
-                  val newStorage = Storage[Float](data)
-                  storages(storageId) = newStorage
-                  newStorage
-                }
-              } else created.asInstanceOf[Storage[Float]]
-              Tensor[Float](storage, offSet, sizes, strides)
-            case TensorType.QUANT => quant()
-          }
-        case DataType.DOUBLE =>
-          val storage : Storage[Double] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getDoubleDataList.asScala.toArray.map(_.doubleValue())
-              val newStorage = Storage[Double](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Double]]
-          Tensor[Double](storage, offSet, sizes, strides)
-        case DataType.BOOL =>
-          val storage : Storage[Boolean] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getBoolDataList.asScala.toArray.map(_.booleanValue())
-              val newStorage = Storage[Boolean](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Boolean]]
-          Tensor[Boolean](storage, offSet, sizes, strides)
-        case DataType.CHAR =>
-          val storage: Storage[Char] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getIntDataList.asScala.toArray.map(_.toChar.charValue())
-              val newStorage = Storage[Char](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Char]]
-          Tensor[Char](storage, offSet, sizes, strides)
-        case DataType.STRING =>
-          val storage: Storage[String] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getStringDataList.asScala.toArray
-              val newStorage = Storage[String](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[String]]
-          Tensor[String](storage, offSet, sizes, strides)
-        case DataType.INT32 =>
-          val storage: Storage[Int] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getIntDataList.asScala.toArray.map(_.intValue())
-              val newStorage = Storage[Int](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Int]]
-          Tensor[Int](storage, offSet, sizes, strides)
-        case DataType.SHORT =>
-          val storage: Storage[Short] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getIntDataList.asScala.toArray.map(_.shortValue())
-              val newStorage = Storage[Short](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Short]]
-          Tensor[Short](storage, offSet, sizes, strides)
-        case DataType.INT64 =>
-          val storage: Storage[Long] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getLongDataList.asScala.toArray.map(_.longValue())
-              val newStorage = Storage[Long](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[Long]]
-          Tensor[Long](storage, offSet, sizes, strides)
-        case DataType.BYTES =>
-          val storage: Storage[ByteString] = if (created == null ) {
-            if (storageId == -1) {
-              null
-            } else {
-              val data = serializedStorage.getBytesDataList.asScala.toArray
-              val newStorage = Storage[ByteString](data)
-              storages(storageId) = newStorage
-              newStorage
-            }
-          } else created.asInstanceOf[Storage[ByteString]]
-          Tensor[ByteString](storage, offSet, sizes, strides)
-        case _ => throw new IllegalArgumentException(s"$dataType not supported in tensor now !")
-      }
-      storages(tensorId) = tensor
-      tensor
-    }
-
-    private def setStorage[T: ClassTag](context: SerializeContext[T],
-      tensorBuilder: BigDLTensor.Builder, tensor: Tensor[_]): Unit = {
-      val storageType = context.storageType
-      if (storageType == ProtoStorageType) {
-        ProtoTensorStorageManager.setStorage(context, tensorBuilder, tensor)
-      } else if (storageType == BigDLStorage) {
-        BigDLTensorStorageManager.setStorage(context, tensorBuilder, tensor)
-      } else {
-        throw new IllegalArgumentException(s"$storageType not supported")
-      }
-    }
-
-    override def setAttributeValue[T: ClassTag]
-      (context: SerializeContext[T], attributeBuilder: AttrValue.Builder, value: Any,
-       valueType : universe.Type = null)
-      (implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.TENSOR)
-      if (value != null) {
-        val tensor = value.asInstanceOf[Tensor[_]]
-        val tensorId = System.identityHashCode(tensor)
-        val storages = context.storages
-        // Check if tensor has been shared
-        if (storages.contains(tensorId)) {
-          attributeBuilder.setTensorValue(resetTensor(storages.get(tensorId).get
-            .asInstanceOf[BigDLTensor]))
-        } else {
-          val totalElement = tensor.nElement()
-          val dimension = tensor.dim()
-          val tensorBuilder = BigDLTensor.newBuilder
-          tensorBuilder.setId(tensorId)
-          tensorBuilder.setDimension(dimension)
-          tensorBuilder.setNElements(totalElement)
-          tensor.getTensorType match {
-            case DenseType =>
-              tensorBuilder.setOffset(tensor.storageOffset())
-              tensorBuilder.setIsScalar(tensor.isScalar)
-              tensorBuilder.setTensorType(TensorType.DENSE)
-            case QuantizedType =>
-              tensorBuilder.setTensorType(TensorType.QUANT)
-          }
-
-          val tensorEmpty = isEmptyTensor(tensor)
-
-          if (!tensorEmpty) {
-            tensor.size().foreach(size => tensorBuilder.addSize(size))
-            tensor.stride().foreach(stride => tensorBuilder.addStride(stride))
-          }
-          setStorage(context, tensorBuilder, tensor)
-          val tensorBuild = tensorBuilder.build
-          attributeBuilder.setTensorValue(resetTensor(tensorBuild))
-          storages(tensorId) = tensorBuild
-        }
-      }
-    }
-
-
-    private def resetStorage(originStorage : TensorStorage) : TensorStorage = {
-      val storageBuilder = TensorStorage.newBuilder
-      storageBuilder.setDatatype(originStorage.getDatatype)
-      storageBuilder.setId(originStorage.getId)
-      storageBuilder.build
-    }
-
-    private def resetTensor(originTensor: BigDLTensor) : BigDLTensor = {
-      val tensorBuilder = BigDLTensor.newBuilder(originTensor)
-      tensorBuilder.clearStorage
-      tensorBuilder.setDatatype(originTensor.getDatatype)
-      tensorBuilder.setId(originTensor.getId)
-      if (originTensor.hasStorage) {
-        tensorBuilder.setStorage(resetStorage(originTensor.getStorage))
-      }
-      tensorBuilder.build
-    }
-  }
-
-/**
- * DataConverter for [[com.intel.analytics.bigdl.nn.VariableFormat]]
- */
-  object VariableFormatConverter extends DataConverter {
-
-    override def getAttributeValue[T: ClassTag](context: DeserializeContext, attribute: AttrValue)
-      (implicit ev: TensorNumeric[T]): AnyRef = {
-      val format = attribute.getVariableFormatValue
-      format match {
-        case VarFormat.DEFAULT => VariableFormat.Default
-        case VarFormat.ONE_D => VariableFormat.ONE_D
-        case VarFormat.IN_OUT => VariableFormat.IN_OUT
-        case VarFormat.OUT_IN => VariableFormat.OUT_IN
-        case VarFormat.IN_OUT_KW_KH => VariableFormat.IN_OUT_KW_KH
-        case VarFormat.OUT_IN_KW_KH => VariableFormat.OUT_IN_KW_KH
-        case VarFormat.GP_OUT_IN_KW_KH => VariableFormat.GP_OUT_IN_KW_KH
-        case VarFormat.GP_IN_OUT_KW_KH => VariableFormat.GP_IN_OUT_KW_KH
-        case VarFormat.OUT_IN_KT_KH_KW => VariableFormat.OUT_IN_KT_KH_KW
-        case VarFormat.EMPTY_FORMAT => null
-      }
-    }
-
-    override def setAttributeValue[T: ClassTag](
-    context: SerializeContext[T], attributeBuilder: AttrValue.Builder,
-      value: Any, valueType: universe.Type = null)(implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.VARIABLE_FORMAT)
-      if (value != null) {
-        val format = value.asInstanceOf[VariableFormat]
-        val formatValue = format match {
-          case VariableFormat.Default => VarFormat.DEFAULT
-          case VariableFormat.ONE_D => VarFormat.ONE_D
-          case VariableFormat.IN_OUT => VarFormat.IN_OUT
-          case VariableFormat.OUT_IN => VarFormat.OUT_IN
-          case VariableFormat.IN_OUT_KW_KH => VarFormat.IN_OUT_KW_KH
-          case VariableFormat.OUT_IN_KW_KH => VarFormat.OUT_IN_KW_KH
-          case VariableFormat.GP_OUT_IN_KW_KH => VarFormat.GP_OUT_IN_KW_KH
-          case VariableFormat.GP_IN_OUT_KW_KH => VarFormat.GP_IN_OUT_KW_KH
-          case VariableFormat.OUT_IN_KT_KH_KW => VarFormat.OUT_IN_KT_KH_KW
-        }
-        attributeBuilder.setVariableFormatValue(formatValue)
-      } else {
-        attributeBuilder.setVariableFormatValue(VarFormat.EMPTY_FORMAT)
-      }
-   }
-}
-/**
- * DataConverter for [[com.intel.analytics.bigdl.nn.InitializationMethod]]
- */
-  object InitMethodConverter extends DataConverter {
-
-    override def getAttributeValue[T: ClassTag](context: DeserializeContext, attribute: AttrValue)
-      (implicit ev: TensorNumeric[T]): AnyRef = {
-      val initMemethod = attribute.getInitMethodValue
-      val initType = initMemethod.getMethodType
-      val methodData = initMemethod.getDataList
-      initType match {
-        case InitMethodType.RANDOM_UNIFORM => RandomUniform
-        case InitMethodType.RANDOM_UNIFORM_PARAM =>
-          RandomUniform(methodData.get(0), methodData.get(1))
-        case InitMethodType.RANDOM_NORMAL =>
-          RandomNormal(methodData.get(0), methodData.get(1))
-        case InitMethodType.ZEROS => Zeros
-        case InitMethodType.ONES => Ones
-        case InitMethodType.CONST => ConstInitMethod(methodData.get(0))
-        case InitMethodType.XAVIER => Xavier
-        case InitMethodType.BILINEARFILLER => BilinearFiller
-        case InitMethodType.EMPTY_INITIALIZATION => null
-      }
-   }
-
-    override def setAttributeValue[T: ClassTag](
-    context: SerializeContext[T], attributeBuilder: AttrValue.Builder,
-      value: Any, valueType: universe.Type = null)(implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.INITMETHOD)
-      val initMethodBuilder = InitMethod.newBuilder
-      if (value != null) {
-        val initMethod = value.asInstanceOf[InitializationMethod]
-        initMethod match {
-          case RandomUniform =>
-            initMethodBuilder.setMethodType(InitMethodType.RANDOM_UNIFORM)
-          case ru: RandomUniform =>
-            initMethodBuilder.setMethodType(InitMethodType.RANDOM_UNIFORM_PARAM)
-            initMethodBuilder.addData(ru.lower)
-            initMethodBuilder.addData(ru.upper)
-          case rm: RandomNormal =>
-            initMethodBuilder.setMethodType(InitMethodType.RANDOM_NORMAL)
-            initMethodBuilder.addData(rm.mean)
-            initMethodBuilder.addData(rm.stdv)
-          case Zeros =>
-            initMethodBuilder.setMethodType(InitMethodType.ZEROS)
-          case Ones =>
-            initMethodBuilder.setMethodType(InitMethodType.ONES)
-          case const: ConstInitMethod =>
-            initMethodBuilder.setMethodType(InitMethodType.CONST)
-            initMethodBuilder.addData(const.value)
-          case Xavier =>
-            initMethodBuilder.setMethodType(InitMethodType.XAVIER)
-          case BilinearFiller =>
-            initMethodBuilder.setMethodType(InitMethodType.BILINEARFILLER)
-        }
-        attributeBuilder.setInitMethodValue(initMethodBuilder.build)
-      } else {
-        initMethodBuilder.setMethodType(InitMethodType.EMPTY_INITIALIZATION)
-        attributeBuilder.setInitMethodValue(initMethodBuilder.build)
-      }
-    }
-  }
-
-  /**
-   * DataConverter for [[com.intel.analytics.bigdl.nn.abstractnn.DataFormat]]
-   */
-  object DataFormatConverter extends DataConverter {
-    override def getAttributeValue[T: ClassTag](context: DeserializeContext, attribute: AttrValue)
-      (implicit ev: TensorNumeric[T]): AnyRef = {
-      val dataFormat = attribute.getDataFormatValue
-      dataFormat match {
-        case InputDataFormat.NCHW => NCHW
-        case InputDataFormat.NHWC => NHWC
-      }
-
-    }
-
-    override def setAttributeValue[T: ClassTag]
-    (context: SerializeContext[T],
-     attributeBuilder: AttrValue.Builder, value: Any, valueType: universe.Type)
-    (implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.DATA_FORMAT)
-      if (value != null) {
-        val dataFormat = value.asInstanceOf[DataFormat]
-        val inputFormat = dataFormat match {
-          case NCHW => InputDataFormat.NCHW
-          case NHWC => InputDataFormat.NHWC
-        }
-        attributeBuilder.setDataFormatValue(inputFormat)
-      }
-    }
-  }
-
-/**
- * DataConverter for [[com.intel.analytics.bigdl.nn.abstractnn.AbstractModule]]
- */
-  object ModuleConverter extends DataConverter {
-
-    override def getAttributeValue[T: ClassTag](context: DeserializeContext, attribute: AttrValue)
-     (implicit ev: TensorNumeric[T]): AnyRef = {
-     val serializedModule = attribute.getBigDLModuleValue
-      if (serializedModule.getModuleType != null && serializedModule.getModuleType != "") {
-        ModuleSerializer.load(DeserializeContext(serializedModule,
-            context.storages, context.storageType)).module
-      } else {
-        null
-      }
-   }
-
-    override def setAttributeValue[T: ClassTag](context: SerializeContext[T],
-      attributeBuilder: AttrValue.Builder,
-      value: Any, valueType: universe.Type = null)(implicit ev: TensorNumeric[T]): Unit = {
-      attributeBuilder.setDataType(DataType.MODULE)
-      if (value != null) {
-        val module = value.asInstanceOf[AbstractModule[Activity, Activity, T]]
-        val serializableModule = ModuleSerializer.
-            serialize(SerializeContext(ModuleData(module, Seq[String](), Seq[String]()),
-              context.storages, context.storageType)).bigDLModule
-          attributeBuilder.setBigDLModuleValue(serializableModule)
-      }
     }
   }
