+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/serializer/converters/DataReaderWriter.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.utils.serializer
+package com.intel.analytics.bigdl.utils.serializer.converters
 
 import java.io.{DataInputStream, DataOutputStream}
